+++ packages/contracts/contracts/Migrations.sol
@@ -14,7 +14,7 @@ contract Migrations {
     owner = msg.sender;
   }
 
-  function setCompleted(uint completed) public restricted {
+  function setCompleted(uint256 completed) public restricted {
     last_completed_migration = completed;
   }
